✨ add Expection Handling

+++ src/socket/Client.java
@@ -2,7 +2,7 @@ package socket;
 
 import java.io.BufferedReader;
 import java.io.DataInputStream;
-import java.io.DataOutputStream;
+import java.io.File;
 import java.io.FileOutputStream;
 import java.io.InputStreamReader;
 import java.io.PrintWriter;
@@ -10,8 +10,10 @@ import java.net.ConnectException;
 import java.net.Socket;
 import java.util.Scanner;
 
+import socket.exceptions.FileCheckException;
+
 public class Client {
-    private static DataOutputStream dataOutputStream = null;
+    static FileOutputStream fileOutputStream;
     private static DataInputStream dataInputStream = null;
     private static int port = 5000;
     private static String ip = "localhost";
@@ -24,41 +26,61 @@ public class Client {
             dataInputStream = new DataInputStream(socket.getInputStream());
 
             while (true) {
+
                 String response;
                 while ((response = in.readLine()) != null) {
                     System.out.println(response);
-                    if(response.equals("Insert Path to File:")) {
+                    if (response.equals("Insert File Index:")) {
                         String input = terminalInput.nextLine();
                         out.println(input);
-                        out.flush();
-                        receiveFile("./src/downloads/" + input);
-                    }
-                    if(!in.ready()) break;
+                        String filePath = in.readLine();
 
+                        if (filePath != null)
+                            receiveFile("./src/downloads/" + filePath);
+                        else {
+                            socket.close();
+                            terminalInput.close();
+                            if (fileOutputStream != null)
+                                fileOutputStream.close();
+                            throw new FileCheckException("File Not Identified");
+                        }
+                    }
+
+                    if (!in.ready())
+                        break;
                 }
-                
+
                 String input = terminalInput.nextLine();
-                if(input.equals("4")) {
+                if (input.equals("4")) {
                     socket.close();
                     terminalInput.close();
+                    if (fileOutputStream != null)
+                        fileOutputStream.close();
                     return;
-                };
-                
+                }
+                ;
+
                 out.println(input);
-                out.flush();
+
             }
         } catch (ConnectException e) {
             System.out.println("Couldn't connect to " + ip + ":" + port);
+        } catch (FileCheckException e) {
+            System.out.println(e.getMessage());
         } catch (Exception e) {
-
             e.printStackTrace();
         }
     }
 
     private static void receiveFile(String fileName) throws Exception {
-        int bytes = 0;
-        FileOutputStream fileOutputStream = new FileOutputStream(fileName);
+        if (fileName.equals("./src/downloads/") || fileName.equals("./src/downloads/null"))
+            throw new FileCheckException("File Not Valid");
+        File f = new File(fileName);
+        if (f.exists() && !f.isDirectory())
+            throw new FileCheckException("File already exists");
 
+        int bytes = 0;
+        fileOutputStream = new FileOutputStream(fileName);
         long size = dataInputStream.readLong();
         byte[] buffer = new byte[4 * 1024];
         while (size > 0 && (bytes = dataInputStream.read(buffer, 0, (int) Math.min(buffer.length, size))) != -1) {
